[Upd] moved registration view to users app
